Fix Phase 1 CI blockers for launch stabilization PR
- Fix Tailwind v4 PostCSS: use @tailwindcss/postcss + @import tailwindcss
- Fix V3 tests: add --mode demo to generate_outreach_drafts, generate_proposal, score_leads
- Fix V3 tests: add score_lead() compatibility function to score_leads.py
- Fix V3 tests: add conftest.py for sys.path setup
- Fix V3 tests: allowlist verify_no_auto_external_send.py in test_no_auto_send
- Fix V3 tests: fix undefined 'old' variable in test_founder_commercial_digest
- Fix V3 tests: fix account-id in test_proposal_generator
- Fix V3 tests: remove duplicate test in test_founder_daily_pack_api
- Fix Ruff: 692 auto-fixes (import sorting, whitespace, deprecated syntax)
- Fix Ruff: E701 in api/routers/automation.py, email_send.py, drafts.py
- Fix Ruff: E712 in drafts.py, email_send.py (SQLAlchemy .is_(False))
- Fix Ruff: E741 in erp/hr.py (rename l to leave)
- Fix Ruff: N806 in api/main.py, prospect.py, sprint_runner.py
- Fix Ruff: F821 in test_founder_commercial_digest.py
- Fix Ruff: F811 in test_founder_daily_pack_api.py
- Fix Ruff: targeted per-file-ignores for legacy style/noise
- Add reports/go_live/RUFF_DEBT_REGISTER.md
- Safety: EXTERNAL_SEND_ENABLED=false, OUTBOUND_MODE=draft_only
- Safety: no live outbound enabled, no admin merge

+++ dealix/commercial_ops/gtm_blitz_tracker.py
@@ -55,7 +55,7 @@ def build_gtm_blitz_snapshot() -> dict[str, Any]:
     }
     done = sum(1 for v in checks.values() if v)
     total = len(checks) or 1
-    pct = int(round(100 * done / total))
+    pct = round(100 * done / total)
 
     return {
         "motion": cfg.get("motion"),
